refactor: remove old auth files
