--- packages/web/src/components/pages/auth/AuthSide/Login/DerivableAccounts/index.ts
@@ -1,1 +1,0 @@
-export { DerivableAccounts } from './DerivableAccounts';
