--- packages/web/src/components/pages/auth/AuthSide/Login/Main/index.ts
@@ -1,1 +1,0 @@
-export { Main } from './Main';
