--- packages/web/src/components/pages/auth/AuthSide/Login/Restore/SelectorAccountItem/index.ts
@@ -1,1 +1,0 @@
-export { SelectorAccountItem } from './SelectorAccountItem';
