--- packages/web/src/components/pages/auth/AuthSide/Login/Restore/index.ts
@@ -1,1 +1,0 @@
-export { Restore } from './Restore';
