--- packages/web/src/components/pages/auth/AuthSide/Login/index.ts
@@ -1,1 +1,0 @@
-export { Login } from './Login';
