--- packages/web/src/components/pages/auth/AuthSide/Ready/OffPasswordModal/index.ts
@@ -1,1 +1,0 @@
-export { OffPasswordModal } from './OffPasswordModal';
