--- packages/web/src/components/pages/auth/AuthSide/Ready/index.ts
@@ -1,1 +1,0 @@
-export { Ready } from './Ready';
