--- packages/web/src/components/pages/auth/AuthSide/Signup/Mnemonic/index.ts
@@ -1,1 +1,0 @@
-export { Mnemonic } from './Mnemonic';
