--- packages/web/src/components/pages/auth/AuthSide/Signup/Paste/index.ts
@@ -1,1 +1,0 @@
-export { Paste } from './Paste';
