--- packages/web/src/components/pages/auth/AuthSide/Signup/index.ts
@@ -1,1 +1,0 @@
-export { Signup } from './Signup';
