--- packages/web/src/components/pages/auth/AuthSide/common/Back/index.ts
@@ -1,1 +1,0 @@
-export { Back } from './Back';
