--- packages/web/src/components/pages/auth/AuthSide/common/Button/index.ts
@@ -1,1 +1,0 @@
-export { Button } from './Button';
