--- packages/web/src/components/pages/auth/AuthSide/common/Password/index.ts
@@ -1,1 +1,0 @@
-export { Password } from './Password';
