--- packages/web/src/components/pages/auth/AuthSide/common/Selector/index.ts
@@ -1,1 +1,0 @@
-export { Selector } from './Selector';
