--- packages/web/src/components/pages/auth/AuthSide/index.ts
@@ -1,1 +1,0 @@
-export { AuthSide } from './AuthSide';
