+++ packages/web/src/new/scenes/Main/Auth/Subviews/components/assets/app.png
