+++ packages/web/src/pages/Landing.tsx
@@ -1,5 +1,5 @@
 import type { FC } from 'react';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { Helmet } from 'react-helmet';
 
 import { css } from '@linaria/core';
@@ -13,9 +13,7 @@ import { Header } from 'components/pages/landing/Header';
 import { fonts } from 'components/pages/landing/styles/fonts';
 import { Top } from 'components/pages/landing/Top';
 import { Under } from 'components/pages/landing/Under';
-// import { Updates } from 'components/pages/landing/Updates';
 import { YouCan } from 'components/pages/landing/YouCan/YouCan';
-import { trackEvent } from 'utils/analytics';
 
 const Wrapper = styled.div`
   display: flex;
@@ -35,10 +33,6 @@ export const global = css`
 
 export const Landing: FC = () => {
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
-
-  useEffect(() => {
-    trackEvent('landing_open');
-  }, []);
 
   const handleDrawerToggle = () => {
     setIsDrawerOpen(!isDrawerOpen);
